pyproject.toml
[tool.poetry]
name = "de-case-response"
version = "0.1.0"
description = ""
authors = ["Alecsander Guimaraes <alecsandergr6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pandas = "^2.1.1"
fastparquet = "^2023.8.0"
pyarrow = "^13.0.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
